Revert "updated tests to cover 95% of code"
