+++ package.json
@@ -22,7 +22,7 @@
     "server": "nodemon server.js",
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --coverage --watchAll=false",
+    "test": "react-scripts test",
     "eject": "react-scripts eject",
     "format": "prettier --write src"
   },
@@ -39,15 +39,6 @@
       "last 1 chrome version",
       "last 1 firefox version",
       "last 1 safari version"
-    ]
-  },
-  "jest": {
-    "collectCoverageFrom": [
-      "**/*.{js,jsx}",
-      "!**/node_modules/**",
-      "!**/*serviceWorker.js",
-      "!**/*index.js",
-      "!**/utils/**"
     ]
   }
 }
